requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
openai==1.23.4
python-dotenv==1.0.1
streamlit==1.33.0
black==24.4.1